requirements.txt
# tensorflow-cpu==2.8.3
gunicorn==20.1.0
uvicorn==0.18.3
# numpy==1.20.3
# matplotlib==3.4.3
fastapi==0.85.1
pillow==8.4.0
python-multipart
tflite-runtime=2.10.0